+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 ### CHANGELOG
 
+#### v0.20.1 2025-11-27
+
+- Added a `collection_list` method.
+- Added a `delete_collection` method.
+- The `napalm` method is now synchronous.
+- Updated the `_Meta` class to store collection metadata.
+- Updated tests.
+- Updated documentation.
+- Updated README.md
+
 #### v0.16.1 2025-10-15
 
 - Added support `python 3.14`.